backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
Flask-Cors
PyJWT
python-dotenv
werkzeug